+++ src/Karl.Abstractions/Models/IAttachmentSource.cs
@@ -1,0 +1,8 @@
+namespace Karl.Models;
+
+public interface IAttachmentSource
+{
+    bool OwnsStream { get; }
+
+    Task<Stream> OpenReadAsync(CancellationToken cancellationToken = default);
+}